+++ CHANGELOG.md
@@ -5,3 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+
+## Added
+
+- Basic Notecard driver support over I2C and UART
+- Two basic samples, one for the each of the communication busses, that request
+  version string from notecard.